Add basic typed wrapper around matching algorithm
This will implement a wrapper around the matching algorithm that is simple and clean to use (at least I hope this). It tries to be as general as possible by using the terms "Helper" and "Helpee". Basic tests are also implemented to ensure everything works (at least a little bit) as expected 😉
A still known issue is that the stats' output format is, in my opinion, more complex than required.  This way, the stats aren't really implemented in the wrapper, so use them with caution.