Rename "dissolved" matches to "exclude" matches
The names used in the Typescript wrapper are renamed to dismiss the word "dissolved", because this could be misleading as we're allowing helpers and helpees to have an open match request count greater than 1 (as we want our code to be as much as general as possible). This means we wanna have a way for forbidding some matches to be made again, cause it's possible for both partner of a match to be part of a matching again. Thus we're excluding some people from being matched again with certain persons.